Adds automation of OWL format generation and deployment
This makes the assumption that the primary ontologies for which OWL format
generation and deployment to web is desired are developed in OWL Functional
Syntax format (.ofn). Hence, it won't do anything for ontologies (such as
demo ontology files etc) that aren't in .ofn format, and it will at present
only act on .ofn files in the root directory of the repo.

The main tool that is used for automatic annotation and conversion is
[robot]. It needs to be installed before this Makefile will work. By
default it is assumed to be in the ./bin directory, but the location
can easily be overridden by `make ROBOT="/path/to/robot"`.

Also provides targets for profile validation. This also depends on robot.

[robot]: http://robot.obolibrary.org/