Removes annotations on :TUMatch class, which is currently obsolete
We were considering the need for a :TUMatch class (see #9), but have for
now decided that there isn't a clear supporting use case. Apparently the
class was once created with a definition and label, and then the class
declaration but not the annotation properties were removed? This does the
latter.